Use a dedicated type for translations in the TranslationTable
Instead of just adding the translations to a hashmap use a dedicated
type (essentially a triplet), containing the from and to strings plus
a set of constraints that apply to the translation, i.e. whether it
only applies at the beginning of a word or only at the end, etc.

This can the be used when finding a matching translation. Instead of
just looking for the longest match we will look for the longest match
that also satisfies the constraints.