Expose the translation in the cli
For that I made the repl function more generic. It accepts a function
now that handles one line input. For translation this is even a
closure that closes over the translation table.

I'm not entirely sure if the mental load that I'm introducing here is
really worth it. The whole thing could be implemented much simpler
with a bit of code duplication