Drop the TranslationMapping type and use Cow instead
In most cases we just need a reference to a Translation in the
translation table. In some cases we do have to create a new
Translation, e.g. when a character is only defined with the
`undefined` rule or if it isn't defined altogether.

Perfect use case for std::bowworw::Cow