Add a second compile pass to resolve implicit dots
It sort of works, unfortunately it only resolves implicit dots within
one table (before expanding the includes). In other words it doesn't
work for any table that includes character definitions, which is
basically most of them.

The implicit resolving code probably needs to be moved to the
table_expanded function where all rules are known. But then we need to
extend it somewhat to pass the path and the line of the rule along