Sketch out an implementation of the match opcode
Unfortunately the proposed implementation matches the characters
including the pre and post pattern. It has currently no way to tell
which one is which. In theory we would need look-ahead and look-behind
to implement the pre and post patterns. Whether this can be handled
with the current simplistic approach remains to be seen.